Instances

+++ EstablecimientoComercial/src/Modelos/Cliente.java
@@ -7,7 +7,7 @@ package Modelos;
 public class Cliente {
     private String nombre;
     private int identificacion;
-    private int telefono;
+    private long telefono;
     private String email;
     private int puntos;
     private double deudas;
@@ -15,7 +15,7 @@ public class Cliente {
     public Cliente(){
         this("Null",0,0,"Null@");
     }
-    public Cliente(String nombre, int identificacion, int telefono, String email) {
+    public Cliente(String nombre, int identificacion, long telefono, String email) {
         this.nombre = nombre;
         this.identificacion = identificacion;
         this.telefono = telefono;
@@ -32,7 +32,7 @@ public class Cliente {
         return identificacion;
     }
 
-    public int getTelefono() {
+    public long getTelefono() {
         return telefono;
     }
 
@@ -56,7 +56,7 @@ public class Cliente {
         this.identificacion = identificacion;
     }
 
-    public void setTelefono(int telefono) {
+    public void setTelefono(long telefono) {
         this.telefono = telefono;
     }
